+++ fabric/src/main/java/com/alien/fabric/data/recipe/impl/MiscellaneousRecipeProvider.java
@@ -33,6 +33,19 @@ public class MiscellaneousRecipeProvider {
             .requires(1, AlienBlocks.ROYAL_JELLY_BLOCK)
             .into(9, AlienItems.RAW_ROYAL_JELLY);
 
+        builder.shaped()
+            .withCategory(RecipeCategory.MISC)
+            .define('I', AlienItems.RAW_SCOURGE_JELLY)
+            .pattern("III")
+            .pattern("III")
+            .pattern("III")
+            .into(1, AlienBlocks.SCOURGE_JELLY_BLOCK);
+
+        builder.shapeless()
+            .withCategory(RecipeCategory.MISC)
+            .requires(1, AlienBlocks.SCOURGE_JELLY_BLOCK)
+            .into(9, AlienItems.RAW_SCOURGE_JELLY);
+
         builder.shapeless()
             .withCategory(RecipeCategory.MISC)
             .requires(1, Items.POISONOUS_POTATO)