+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ## ✨ What's New
 - Added raw scourge jelly item.
+- Added scourge jelly block.
 - Added a new set of xenomorphs... the SCOURGE xenomorphs:
   - Added chrysalis.
     - Can roll around in a ball to cross distances quickly.